# contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Internal Audience & Campaign Control API
  version: 0.8.28
  description: >
    Internal-only API for organisation records, offline consent reviews, secure audience imports,
    governed cohort definitions and controlled campaign lifecycle management.
servers:
  - url: /api/v1
security:
  - internalSession: []
paths:
...
  /admin/provider-capabilities:
    get:
      operationId: listProviderCapabilities
      summary: List governed provider and engine capability definitions
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque continuation token returned by the previous page. }
      responses:
        '200':
          description: Provider capability definition list
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
    post:
      operationId: createProviderCapability
      summary: Create a draft provider capability definition
      responses:
        '201': { description: Draft provider capability definition created }
  /admin/provider-capabilities/{id}/submit:
    post:
      operationId: submitProviderCapability
      summary: Submit a provider capability definition for independent approval
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Provider capability definition awaiting approval }
  /admin/provider-capabilities/{id}/decision:
    post:
      operationId: decideProviderCapability
      summary: Approve or reject a provider capability definition
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Provider capability decision recorded }
        '403': { description: Recent MFA step-up required }
  /admin/provider-capabilities/{id}/retire:
    post:
      operationId: retireProviderCapability
      summary: Retire an active provider capability definition
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Provider capability definition retired }
        '403': { description: Recent MFA step-up required }
  /admin/provider-capabilities/{id}/events:
    get:
      operationId: listProviderCapabilityEvents
      summary: List append-only provider capability lifecycle events
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
        - { name: cursor, in: query, required: false, schema: { type: string }, description: Opaque cursor returned by the previous page. }
      responses:
        '200':
          description: Provider capability event history
          content:
            application/json:
              schema:
                type: object
                required: [items, count, hasMore]
                properties:
                  items: { type: array, items: { type: object } }
                  count: { type: integer, minimum: 0 }
                  nextCursor: { type: string }
                  hasMore: { type: boolean }
        '400': { description: Invalid page limit or cursor }
  /admin/provider-capabilities/resolve:
    post:
      operationId: resolveProviderCapability
      summary: Resolve an active provider and engine definition and verify required capabilities
      responses:
        '200': { description: Active compatible provider capability definition }
        '404': { description: No active provider capability definition matched }
        '422': { description: Required capabilities are not supported }
...
components:
  securitySchemes:
    internalSession:
      type: apiKey
      in: cookie
      name: campaign_session
...